Consolidate scattered shared files into shared/ directory
- Move root components/ (2 files) → shared/components/
- Move root services/ (2 files) → shared/services/
- Move root stores/ (4 files) → shared/stores/
- Move root utils/ (3 files) → shared/utils/
- Update all import paths throughout codebase (50+ files)
- Clean up root directory structure
- Fix all remaining utils import path issues

Major cleanup: Eliminates ambiguous root-level directories and
consolidates all truly shared code in one predictable location.
