+++ ui/src/navigation/components/ShopView.svelte
@@ -16,7 +16,7 @@
     searchResultsStore,
     searchMethodStore,
     featuredSubcategories,
-  } from "../../stores/UiOnlyStore";
+  } from "../../shared/stores/UiOnlyStore";
 
   // Import from data trigger store (keeping sort/filter stores for now)
   // Note: searchMode and searchQuery now come from DataManager navigationState
@@ -24,7 +24,7 @@
   // No longer need DataManager context - using direct imports
 
   // Import NavigationStore and category utilities
-  import { navigationStore } from "../../stores/NavigationStore";
+  import { navigationStore } from "../../shared/stores/NavigationStore";
   import { mainCategories } from "../../products/utils/categoryData";
 
   // Store context removed - using direct service access
