+++ ui/src/products/components/AllProductsGrid.svelte
@@ -3,7 +3,7 @@
     import ProductCard from "./ProductCard.svelte";
     import { createEventDispatcher } from "svelte";
     import SortFilterDropdown from "../../shared/components/SortFilterDropdown.svelte";
-    import { filterState, getSortedFilteredProducts, setSortBy, setSelectedBrands, setSelectedOrganic } from "../../stores/SortingStore";
+    import { filterState, getSortedFilteredProducts, setSortBy, setSelectedBrands, setSelectedOrganic } from "../../shared/stores/SortingStore";
 
 
     import { mainCategories } from "../utils/categoryData";
