+++ ui/src/products/utils/SimpleCloneCache.ts
@@ -4,7 +4,7 @@
 
 import type { AppClient, CellId } from "@holochain/client";
 import { getActiveCloneCellId } from "./cloneHelpers";
-import { startCloneSetup, updateCloneSetup, finishCloneSetup } from "../../stores/LoadingStore";
+import { startCloneSetup, updateCloneSetup, finishCloneSetup } from "../../shared/stores/LoadingStore";
 
 export class SimpleCloneCache {
     private cachedCellId: CellId | null = null;
