+++ ui/src/shared/stores/NavigationStore.ts
