+++ ui/src/shared/utils/animationUtils.ts
