+++ ui/src/shared/utils/priceUtils.ts
