+++ ui/src/shared/utils/stockUtils.ts
